+++ classes/unplag_core.class.php
@@ -20,6 +20,7 @@ use core\event\base;
 use plagiarism_unplag;
 use plagiarism_unplag\classes\entities\unplag_archive;
 use plagiarism_unplag\classes\exception\unplag_exception;
+use plagiarism_unplag\classes\helpers\unplag_check_helper;
 use plagiarism_unplag\classes\helpers\unplag_progress;
 use plagiarism_unplag\classes\plagiarism\unplag_file;
 
@@ -116,7 +117,7 @@ class unplag_core {
                     throw new unplag_exception($resp->errors);
                 }
 
-                self::check_complete($record, $resp->check);
+                unplag_check_helper::check_complete($record, $resp->check);
             } else {
                 $DB->update_record(UNPLAG_FILES_TABLE, $record);
             }
@@ -142,64 +143,6 @@ class unplag_core {
         }
 
         return $record;
-    }
-
-    /**
-     * @param \stdClass $record
-     * @param \stdClass $check
-     * @param int $progress
-     */
-    public static function check_complete(\stdClass &$record, \stdClass $check, $progress = 100) {
-        global $DB;
-
-        if ($progress == 100) {
-            $record->statuscode = UNPLAG_STATUSCODE_PROCESSED;
-        }
-
-        $record->similarityscore = $check->report->similarity;
-        $record->reporturl = $check->report->view_url;
-        $record->reportediturl = $check->report->view_edit_url;
-        $record->progress = $progress;
-
-        $updated = $DB->update_record(UNPLAG_FILES_TABLE, $record);
-
-        $emailstudents = unplag_settings::get_assign_settings($record->cm, 'unplag_studentemail');
-        if ($updated && !empty($emailstudents)) {
-            unplag_notification::send_student_email_notification($record);
-        }
-
-        if ($updated && $record->parent_id !== null) {
-            $parentrecord = $DB->get_record(UNPLAG_FILES_TABLE, array('id' => $record->parent_id));
-            $childs = $DB->get_records(UNPLAG_FILES_TABLE, array('parent_id' => $parentrecord->id, 'errorresponse' => null));
-            $similarity = 0;
-            $cpf = null;
-            $parentprogress = 0;
-            foreach ($childs as $child) {
-                if ($cpf === null) {
-                    $cpf = $child->id;
-                }
-                $parentprogress += $child->progress;
-                $similarity += $child->similarityscore;
-            }
-
-            $parentprogress = floor($parentprogress / count($childs));
-            $reporturl = new \moodle_url('/plagiarism/unplag/reports.php', array(
-                    'cmid' => $parentrecord->cm,
-                    'pf' => $parentrecord->id,
-                    'cpf' => $cpf
-            ));
-
-            $parentcheck = array(
-                    'report' => array(
-                            'similarity' => floor($similarity / count($childs)),
-                            'view_url' => (string) $reporturl->out_as_local_url(),
-                            'view_edit_url' => (string) $reporturl->out_as_local_url()
-                    )
-            );
-
-            $parentcheck = json_decode(json_encode($parentcheck));
-            self::check_complete($parentrecord, $parentcheck, $parentprogress);
-        }
     }
 
     /**
